Split out visualizerUtils
This is partially an attempt to allow me to focus on the Visualizer component so I can convert it to a functional component, but also an attempt to iteratively type this code
